feat: 添加大文件加载确认对话框和流式播放支持
添加加载确认对话框，当打开大文件时提示用户启用内存优化模式。实现基于缓存的流式播放功能，减少大文件的内存占用。重构相关代码以支持新的加载流程和播放机制。

主要变更：
- 新增 load_confirm_dialog 模块和对话框状态管理
- 实现 MidiCache 流式读取和播放引擎集成
- 修改文件加载逻辑以支持内存优化模式
- 更新相关测试和依赖配置

+++ crates/cache/src/chunk.rs
@@ -379,7 +379,7 @@ pub fn phase2_assemble_to_path(
 
                         let sf = shared_file.clone();
                         let mut f = sf.lock().unwrap();
-                        let offset = f.seek(SeekFrom::End(0)).expect("seek") as u64;
+                        let offset = f.seek(SeekFrom::End(0)).expect("seek");
                         f.write_all(&bytes).expect("write");
                         drop(f);
 
@@ -486,7 +486,7 @@ pub fn phase2_assemble<W: Write>(
 
                         let sf = shared_file.clone();
                         let mut f = sf.lock().unwrap();
-                        let offset = f.seek(SeekFrom::End(0)).expect("seek") as u64;
+                        let offset = f.seek(SeekFrom::End(0)).expect("seek");
                         f.write_all(&bytes).expect("write");
                         drop(f);
